refactor: turn defaults into class methods
