+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetUtils.kt
@@ -2,7 +2,6 @@ package xyz.lbres.kotlinutils.set.multiset
 
 import xyz.lbres.kotlinutils.collection.ext.toMultiSet
 import xyz.lbres.kotlinutils.collection.ext.toMutableMultiSet
-import kotlin.math.min
 
 /**
  * Create a MultiSet containing the given elements.
@@ -44,59 +43,3 @@ fun <E> MultiSet(size: Int, init: (Int) -> E): MultiSet<E> = MultiSetImpl((0 unt
  * @return [MutableMultiSet]<E>
  */
 fun <E> MutableMultiSet(size: Int, init: (Int) -> E): MutableMultiSet<E> = MutableMultiSetImpl((0 until size).map(init))
-
-/**
- * Plus function that can be used in any MultiSet implementation.
- * May not be as efficient as implementation-specific code.
- *
- * @param set [MultiSet]<E>: first MultiSet
- * @param other [MultiSet]<E>: second MultiSet
- * @return [MultiSet]<E>: MultiSet containing all elements in both sets
- */
-fun <E> defaultPlus(set: MultiSet<E>, other: MultiSet<E>): MultiSet<E> {
-    val values = (set.distinctValues + other.distinctValues).flatMap { element ->
-        val totalCount = set.getCountOf(element) + other.getCountOf(element)
-        List(totalCount) { element }
-    }
-
-    return values.toMultiSet()
-}
-
-/**
- * Minus function that can be used in any MultiSet implementation.
- * May not be as efficient as implementation-specific code.
- *
- * @param set [MultiSet]<E>: MultiSet to subtract from
- * @param other [MultiSet]<E>: values to subtract from [set]
- * @return [MultiSet]<E>: MultiSet containing the elements in [set] but not in [other]
- */
-fun <E> defaultMinus(set: MultiSet<E>, other: MultiSet<E>): MultiSet<E> {
-    val values = set.distinctValues.flatMap { element ->
-        val totalCount = set.getCountOf(element) - other.getCountOf(element)
-
-        if (totalCount <= 0) {
-            emptyList()
-        } else {
-            List(totalCount) { element }
-        }
-    }
-
-    return values.toMultiSet()
-}
-
-/**
- * Intersect function that can be used in any MultiSet implementation.
- * May not be as efficient as implementation-specific code.
- *
- * @param set [MultiSet]<E>: first MultiSet
- * @param other [MultiSet]<E>: second MultiSet
- * @return [MultiSet]<E>: MultiSet containing only the elements that occur in both sets
- */
-fun <E> defaultIntersect(set: MultiSet<E>, other: MultiSet<E>): MultiSet<E> {
-    val values = (set.distinctValues intersect other.distinctValues).flatMap { element ->
-        val totalCount = min(set.getCountOf(element), other.getCountOf(element))
-        List(totalCount) { element }
-    }
-
-    return values.toMultiSet()
-}
