+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetInlineMethodsTest.kt
@@ -1,9 +1,9 @@
 package xyz.lbres.kotlinutils.set.multiset
 
 import org.junit.Test
-import xyz.lbres.kotlinutils.set.multiset.methods.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.inline.* // ktlint-disable no-wildcard-imports no-unused-imports
 
-internal class MultiSetMethodsTest {
+internal class MultiSetInlineMethodsTest {
     @Test fun testMap() = runMapTests()
     @Test fun testFilter() = runFilterTests()
     @Test fun testFilterNot() = runFilterNotTests()
