+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/inline/runMinMaxTests.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.kotlinutils.set.multiset.methods
+package xyz.lbres.kotlinutils.set.multiset.inline
 
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.assertEquals
